fixed bug in guest cart

+++ src/routes/components/CartForm.js
@@ -11,20 +11,19 @@ const CartForm = ({
   title,
   author,
 }) => {
-  const [bookQuantity, setBookQuantity] = useState(1);
-  //const [guestCart, setGuestCart] = useState([])
+  const [bookQuantity, setBookQuantity] = useState();
+
 
   const addToCartSubmitHandler = async (event) => {
     event.preventDefault();
+    console.log("userId", userId)
 
     if (userId) {
       await addBookToCart(userId, price, id, bookQuantity);
       alert("Book added to cart");
     } else {
       let existingEntries = JSON.parse(localStorage.getItem("GuestCartData"));
-      if (existingEntries == null) {
-        existingEntries = [];
-      }
+
       let newBook = {
         id,
         title,
@@ -34,10 +33,28 @@ const CartForm = ({
         price,
         bookQuantity,
       };
+      console.log("exsitingEntries", existingEntries)
 
-      localStorage.setItem("newBook", JSON.stringify(newBook));
-      existingEntries.push(newBook);
-      localStorage.setItem("GuestCartData", JSON.stringify(existingEntries));
+      if (existingEntries.length < 1) {
+        console.log("sarah")
+        existingEntries = [];
+        existingEntries.push(newBook);
+        localStorage.setItem("GuestCartData", JSON.stringify(existingEntries));
+      } else {
+        const checkForBook = existingEntries.filter((book) => {
+          console.log("book.id", book.id);
+          console.log("newBook.id", newBook.id);
+
+          if (book.id === newBook.id) {
+            const newQuantity = newBook.bookQuantity + book.bookQuantity;
+            newBook.bookQuantity = newQuantity;
+          }
+          return book.id !== newBook.id;
+        });
+        checkForBook.push(newBook);
+        localStorage.setItem("GuestCartData", JSON.stringify(checkForBook));
+        console.log("checkForBook", checkForBook);
+      }
     }
   };
